chore: update rollup example
