+++ tauri/examples/rollup/src-tauri/Cargo.toml
@@ -14,7 +14,7 @@ build = "src/build.rs"
 [dependencies]
 serde_json = "1.0"
 serde = { version = "1.0", features = [ "derive" ] }
-tauri = { path = "../../..", features = [ "all-api", "edge" ] }
+tauri = { path = "../../..", features =["all-api", "cli"]}
 
 [target."cfg(windows)".build-dependencies]
 winres = "0.1"
