+++ tauri/examples/rollup/src-tauri/src/main.rs
@@ -14,25 +14,21 @@ struct Reply {
 
 fn main() {
   tauri::AppBuilder::new()
-    .setup(|webview, _source| {
-      let handle = webview.handle();
+    .setup(|webview, _source| async move {
+      let mut webview = webview.clone();
       tauri::event::listen(String::from("js-event"), move |msg| {
         println!("got js-event with message '{:?}'", msg);
         let reply = Reply {
           data: "something else".to_string(),
         };
 
-        tauri::event::emit(
-          &handle,
-          String::from("rust-event"),
-          Some(serde_json::to_string(&reply).unwrap()),
-        )
-        .expect("failed to emit");
+        tauri::event::emit(&mut webview, String::from("rust-event"), Some(reply))
+          .expect("failed to emit");
       });
     })
-    .invoke_handler(|_webview, arg| {
+    .invoke_handler(|mut webview, arg| async move {
       use cmd::Cmd::*;
-      match serde_json::from_str(arg) {
+      match serde_json::from_str(&arg) {
         Err(e) => Err(e.to_string()),
         Ok(command) => {
           match command {
@@ -48,8 +44,8 @@ fn main() {
               // tauri::execute_promise is a helper for APIs that uses the tauri.promisified JS function
               // so you can easily communicate between JS and Rust with promises
               tauri::execute_promise(
-                _webview,
-                move || {
+                &mut webview,
+                async move {
                   println!("{} {:?}", endpoint, body);
                   // perform an async operation here
                   // if the returned value is Ok, the promise will be resolved with its value
@@ -60,6 +56,7 @@ fn main() {
                 callback,
                 error,
               )
+              .await
             }
           }
           Ok(())
