+++ tauri/examples/rollup/src-tauri/tauri.conf.json
@@ -1,59 +1,70 @@
 {
   "build": {
     "distDir": "../public",
-    "devPath": "http://localhost:5000",
-    "beforeDevCommand": "yarn dev",
-    "beforeBuildCommand": "yarn build"
+    "devPath": "../public",
+    "withGlobalTauri": true,
+    "beforeBuildCommand": "yarn build",
+    "beforeDevCommand": "yarn dev"
   },
-  "ctx": {},
   "tauri": {
-    "cli": null,
+    "cli": {
+      "description": "Tauri communication example",
+      "args": [
+        {
+          "short": "c",
+          "name": "config",
+          "takesValue": true,
+          "description": "Config path"
+        },
+        {
+          "short": "t",
+          "name": "theme",
+          "takesValue": true,
+          "description": "App theme",
+          "possibleValues": ["light", "dark", "system"]
+        },
+        {
+          "short": "v",
+          "name": "verbose",
+          "multipleOccurrences": true,
+          "description": "Verbosity level"
+        }
+      ],
+      "subcommands": {
+        "update": {
+          "description": "Updates the app",
+          "args": [
+            {
+              "short": "b",
+              "name": "background",
+              "description": "Update in background"
+            }
+          ]
+        }
+      }
+    },
     "embeddedServer": {
       "active": true
     },
     "bundle": {
       "active": true,
-      "targets": "all",
-      "identifier": "com.tauri.dev",
+      "identifier": "com.tauri.communication",
       "icon": [
         "icons/32x32.png",
         "icons/128x128.png",
         "icons/128x128@2x.png",
         "icons/icon.icns",
         "icons/icon.ico"
-      ],
-      "resources": [],
-      "externalBin": [],
-      "copyright": "",
-      "category": "DeveloperTool",
-      "shortDescription": "",
-      "longDescription": "",
-      "deb": {
-        "depends": [],
-        "useBootstrapper": false
-      },
-      "osx": {
-        "frameworks": [],
-        "minimumSystemVersion": "",
-        "useBootstrapper": false
-      },
-      "exceptionDomain": ""
+      ]
     },
-    "whitelist": {
+    "allowlist": {
       "all": true
     },
     "window": {
-      "title": "Tauri App",
-      "width": 800,
-      "height": 600,
-      "resizable": true,
-      "fullscreen": false
+      "title": "Tauri API Validation"
     },
     "security": {
       "csp": "default-src blob: data: filesystem: ws: http: https: 'unsafe-eval' 'unsafe-inline'"
-    },
-    "edge": {
-      "active": true
     },
     "inliner": {
       "active": true
